+++ packages/core/src/integrations/ghtml/ghtml.constants.ts
@@ -1,0 +1,1 @@
+export const GHTML_PLUGIN_NAME = 'ghtml';