+++ packages/integrations/mdx/src/mdx.constants.ts
@@ -1,0 +1,1 @@
+export const MDX_PLUGIN_NAME = 'MDX';